requirements.txt
asgiref @ file:///home/runner/.cache/pypoetry/artifacts/44/47/8b/eeb8b3236803e39b50f24aec1dd7dc622a47aedd27d0c37ba347271f9b/asgiref-3.5.1-py3-none-any.whl
backports.entry-points-selectable==1.1.1
CacheControl==0.12.10
cachetools==5.3.0
cachy==0.3.0
certifi @ file:///home/runner/.cache/pypoetry/artifacts/58/94/ad/c92c7ef8cba2444f4da61c1c47e27e96bc6b71043e9a175a2fd20e4495/certifi-2023.5.7-py3-none-any.whl
cffi==1.15.0
charset-normalizer @ file:///home/runner/.cache/pypoetry/artifacts/ed/4c/bf/ecbe5d0f99d3e70e82399fb2141338a3791ad1e83475921d81757ffc25/charset_normalizer-3.1.0-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==35.0.0
debugpy==1.2.1
distlib==0.3.3
Django @ file:///home/runner/.cache/pypoetry/artifacts/0a/6e/a7/ad9b81885c41d7edd5f5fd91479a9c6be684b7ad660172f6b1f93c7d42/Django-3.2.13-py3-none-any.whl
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.3.2
future==0.18.2
google-api-core==2.11.0
google-api-python-client @ file:///home/runner/.cache/pypoetry/artifacts/df/a3/4a/7f91441b9aa70fb7707c9b89fec2d063894fb07681d4ca7f6c62ceac07/google_api_python_client-2.87.0-py2.py3-none-any.whl
google-auth @ file:///home/runner/.cache/pypoetry/artifacts/10/c2/bf/33b0e1c486ae541804a8c17552a22e2f7e11892dcb28bd77988ef9abe0/google_auth-2.19.0-py2.py3-none-any.whl
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grpcio @ file:///home/runner/.cache/pypoetry/artifacts/a7/31/a3/1479f2ff2cb2f33335a3681b88f9d2b3bf6995dae868380f9a14bea5f0/grpcio-1.55.0-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
html5lib==1.1
httplib2==0.22.0
idna @ file:///home/runner/.cache/pypoetry/artifacts/0b/15/a9/ce99545393f0334813ed7f8c93c971de7c34006aad4a32241a43170414/idna-3.4-py3-none-any.whl
importlib-metadata==3.10.1
jedi==0.15.2
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.2
numpy @ file:///home/runner/.cache/pypoetry/artifacts/0a/f1/63/4e86e206dd78b5ffa1d0b33299c8275c3e1a2411fac9d63320ce69de5f/numpy-1.22.2-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
oauthlib==3.2.2
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.4.0
pluggy==0.13.1
poetry==1.1.11
poetry-core==1.0.7
protobuf==4.23.1
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyflakes==2.1.1
pylev==1.4.0
pyparsing @ file:///home/runner/.cache/pypoetry/artifacts/95/8e/67/7c07b50a56fd53f9deb19bccf77afe14d18c31018b923b69c69b328d15/pyparsing-3.0.9-py3-none-any.whl
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pytz @ file:///home/runner/.cache/pypoetry/artifacts/2f/64/84/b6333ff5a89076f2f48899fb6dc7e562400c812c962713b90b44ac5160/pytz-2022.1-py2.py3-none-any.whl
requests @ file:///home/runner/.cache/pypoetry/artifacts/e4/f0/02/7282fe6161617431af21eb6aade409ae268ff2368eb949ad3b32dd772d/requests-2.31.0-py3-none-any.whl
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.9
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sqlparse @ file:///home/runner/.cache/pypoetry/artifacts/5c/57/47/582e6c575570333d1c84c495489bc30ed2c7e0fccaabeaa29650cfc543/sqlparse-0.4.2-py3-none-any.whl
toml==0.10.2
tomlkit==0.7.2
typing-extensions==3.7.4.3
ujson==1.35
uritemplate==4.1.1
urllib3 @ file:///home/runner/.cache/pypoetry/artifacts/56/9d/d9/69a2536f9016a06c5e43c698e3d539e0f50fa75bcff9b1b10a0fb25782/urllib3-1.26.15-py2.py3-none-any.whl
virtualenv==20.10.0
webencodings==0.5.1
xds-protos @ file:///home/runner/.cache/pypoetry/artifacts/a5/c5/58/2120d228439eff775ebbc2b6d66242ccb3e26527a7a004768033747084/xds_protos-0.0.11-py3-none-any.whl
yapf==0.31.0
yarg==0.1.9
zipp==3.4.1
